coremark:Stack overflow in ee_printf with DEBUG_CORE=1.
The state string printed in `core_main.c` can be quite large and the default 256-byte buffer in the `barebones/` implementation of `ee_printf` is insufficient. Since this is DEBUG code, the edit is allowed, as it will not impact performance. Additionally, the version of `ee_printf` provided is an example, and will require platform customization.